feat(security): API-key auth, session/frame limits, pydantic validation, subscribe-only tokens, StaticFiles
- LIVEAVATAR_API_KEY protects REST (X-API-Key) and WS (query/header), 401/4401
- LIVEAVATAR_MAX_SESSIONS caps concurrent sessions (429)
- LIVEAVATAR_MAX_WS_FRAME_BYTES drops oversized WS frames (DoS guard)
- CreateSessionBody pydantic model with id validation (422)
- browser LiveKit tokens are subscribe-only (canPublish=False)
- catch-all static route replaced with StaticFiles mount
- CI fixes: dev extra (httpx2), numpy float32 dtype, mypy interpreter default
- 23 new security tests (119 total)

+++ src/liveavatar/preview.py
@@ -79,7 +79,7 @@ def load_wav_mono_16k(path: str) -> tuple[bytes, float]:
 
     audio = np.frombuffer(raw, dtype=np.int16).astype(np.float32)
     if nchannels > 1:
-        audio = audio.reshape(-1, nchannels).mean(axis=1)
+        audio = audio.reshape(-1, nchannels).mean(axis=1).astype(np.float32)
 
     if framerate != 16000:
         duration = len(audio) / framerate
@@ -88,7 +88,7 @@ def load_wav_mono_16k(path: str) -> tuple[bytes, float]:
             np.linspace(0, len(audio) - 1, target_n),
             np.arange(len(audio)),
             audio,
-        )
+        ).astype(np.float32)
 
     return audio.astype(np.int16).tobytes(), nframes / framerate
